requirements.txt
django>=5.0
psycopg2-binary
django-environ
djangorestframework>=3.14